[posix] filesystem: move to platform folder
