+++ cmake/treedata/ios/subdirs.txt
@@ -3,6 +3,7 @@ xbmc/platform/linux/network           platform/linux/network
 xbmc/input/touch                      input/touch
 xbmc/input/touch/generic              input/touch/generic
 xbmc/platform/posix                   posix
+xbmc/platform/posix/filesystem        platform/posix/filesystem
 xbmc/platform/posix/utils             platform/posix/utils
 xbmc/platform/darwin                  platform/darwin
 xbmc/platform/darwin/ios              platform/ios
@@ -11,7 +12,6 @@ xbmc/platform/darwin/osx/network      platform/osx/network
 xbmc/platform/darwin/osx/peripherals  platform/osx/peripherals
 xbmc/platform/darwin/osx/powermanagement platform/darwin/osx/powermanagement
 xbmc/platform/darwin/osx/storage      platform/osx/storage
-xbmc/filesystem/posix                 filesystem/posix
 xbmc/windowing/osx                    windowing/osx
 xbmc/cores/RetroPlayer/process/ios    cores/RetroPlayer/process/ios
 xbmc/cores/VideoPlayer/Process/ios    cores/VideoPlayer/Process/ios
