+++ xbmc/platform/posix/filesystem/PosixDirectory.cpp
