+++ xbmc/platform/posix/filesystem/PosixDirectory.h
