+++ xbmc/platform/posix/filesystem/PosixFile.cpp
