+++ xbmc/platform/posix/filesystem/PosixFile.h
